Make it more symetric by switching ctrl gui/cmd on right side
Also fix issue with enter being accidentally pushed to often. Enter is
a critical key and should not be accidentally pushed as much...! Let's
see how it goes...

Actually, when I come to think of it... Not sure if this layout is any
better. Since on Mac, enter-key is now on the same key as command...!